requirements.txt
streamlit
faster-whisper
transformers
torch
sentence-transformers
faiss-cpu
python-dotenv
